requirements.txt
segment-anything>=1.0
torch>=2.0.0
numpy>=1.24.0
Pillow>=9.0.0
requests>=2.31.0